requirements.txt
Django
django-allauth
django-crispy-forms
django-materialize-css
